requirements.txt
requests
numpy
biopython
pandas
streamlit
st_pages
py3Dmol